altri piccoli passi per riorganizzazione
