+++ coa/src/internal/engine/remaster_plan.go
@@ -14,80 +14,6 @@ import (
 	"time"
 )
 
-// BootloaderRoot definisce dove vengono estratti i bootloader.
-const BootloaderRoot = "/tmp/coa/bootloaders"
-
-// Action rappresenta un singolo blocco "command" nell'array "plan"
-type Action struct {
-	Command         string   `json:"command"`
-	VolID           string   `json:"volid,omitempty"`
-	OutputISO       string   `json:"output_iso,omitempty"`
-	CryptedPassword string   `json:"crypted_password,omitempty"`
-	RunCommand      string   `json:"run_command,omitempty"`
-	Chroot          bool     `json:"chroot,omitempty"` // Supporto per esecuzione in liveroot
-	ExcludeList     string   `json:"exclude_list,omitempty"`
-	BootParams      string   `json:"boot_params,omitempty"` // Parametri dinamici per il bootloader
-	Args            []string `json:"args,omitempty"`
-}
-
-// UserConfig definisce la struttura per la creazione nativa dell'utente live
-type UserConfig struct {
-	Login    string   `json:"login"`
-	Password string   `json:"password"`
-	Gecos    string   `json:"gecos"`
-	Home     string   `json:"home"`
-	Shell    string   `json:"shell"`
-	Groups   []string `json:"groups"`
-}
-
-// FlightPlan è l'oggetto JSON principale inviato al motore oa
-type FlightPlan struct {
-	PathLiveFs      string       `json:"pathLiveFs"`
-	Mode            string       `json:"mode"`
-	Family          string       `json:"family"`
-	InitrdCmd       string       `json:"initrd_cmd"`
-	BootloadersPath string       `json:"bootloaders_path"`
-	Users           []UserConfig `json:"users"`
-	Plan            []Action     `json:"plan"`
-}
-
-// generateExcludeList crea il file .list dinamico per mksquashfs
-func generateExcludeList(mode string) string {
-	outPath := "/tmp/coa/excludes.list"
-	var excludes []string
-
-	excludes = append(excludes,
-		"boot/efi/EFI",
-		"boot/loader/entries/",
-		"etc/fstab",
-		"var/lib/docker/",
-	)
-
-	if mode != "clone" && mode != "crypted" {
-		excludes = append(excludes, "root/*")
-	}
-
-	userList := "/etc/coa/exclusion.list"
-	if _, err := os.Stat(userList); os.IsNotExist(err) {
-		userList = "conf/exclusion.list"
-	}
-
-	if data, err := os.ReadFile(userList); err == nil {
-		lines := strings.Split(string(data), "\n")
-		for _, line := range lines {
-			line = strings.TrimSpace(line)
-			if line != "" && !strings.HasPrefix(line, "#") {
-				excludes = append(excludes, line)
-			}
-		}
-	}
-
-	os.MkdirAll("/tmp/coa", 0755)
-	os.WriteFile(outPath, []byte(strings.Join(excludes, "\n")+"\n"), 0644)
-
-	return outPath
-}
-
 // GeneratePlan costruisce il piano di volo dinamico
 func GeneratePlan(d *distro.Distro, mode string, workPath string) FlightPlan {
 	plan := FlightPlan{
